Entirely rewrite our Cocoa/node event loop integration.
Manually run the helper thread instead of using a CFFileDescriptor.  This seems
to avoid some un-understood unreliability and allows the operation to be a bit
cleaner and for the synthentic event to come from the helper thread.
Additionally this changes the general approach of how when integrate with libuv,
which removes a lot of the hacks needed to non-block run the uv event loop in
hopes of getting the kqueue state up to date.