Initial commit of "Plask 3.0"
This is quite a big change, mostly do to the V8 API changes
and the addition of isolates.  Other changes include typed
arrays which are now part of V8 itself (although the API
is still experimental and has some problems).  The event
loop was rewritten to follow some changes in Node, and
also hopefully some simplifications like removing the dummy
timer.

Additionally Skia was brought up to date, which mostly just
required some changes related to SkDevice/SkBitmap.  This
should be cleaned up a bit still in the future, along with
the possibility of using the GPU support which is in general
where Skia is concentrating their effort.

> process.versions
{ http_parser: '2.2',
  node: '0.11.13',
  v8: '3.25.30',
  uv: '0.11.25',
  zlib: '1.2.3',
  modules: '14',
  openssl: '1.0.1g' }