Make sure to wake up nextEventMatchingMask after drop event.
Drag and drop is a bit tricky, as it goes through the NSApp event
system but isn't considered an NSEvent so is handled internally
and the call does not return, but there could be changes made by
the JavaScript callback, so we want to go around the loop again.

An alternative would be to queue up the drop events, basically
running the callback as a setTimeout 0, but this should be fine.