Handle UNPACK_PREMULTIPLY_ALPHA_WEBGL, always true in Plask.
Plask and Skia work in premultiplied alpha, which the browsers
probably also do internally.  Instead of wasting time doing the
lossy conversion to unpremultiplied alpha, just fix that Plask
always worked in premultiplied textures.  This deviates from the
WebGL spec in terms of the default value, but makes a lot more
sense.