Add a relink() method to MagicProgram.
When you create a MagicProgram it is linked without a chance to
run any external code beforehand.  For things like transform
feedback the varying settings only take effect when the program
is linked.  It is not the prefect solution but this at least
allows the external changes to be made and then the program relinked.