Upgrade to Node v5.2.0 branched as Plask v4.
This adapts to TypedArray changes in V8, specifically the generic
indexing of "external array data" with [] has been dropped and
TypedArrays are supported directly.  The biggest change for Plask
means that raw pixel access on SkCanvas is no longer done directly
on the canvas object but a .pixels property which is just a regular
Uint8ClampedArray TypedArray object.