eventloop: Manually run microtask queue after dispatching UI events.
Not sure exactly the mechanics of this, there are some places in Node
(for example async wrapper) that manually run the queue.  It seems we
need to at least do this for UI events.  A failing test cause was
previously not receiving Object.observe notifications when a change was
made on the object during a UI event callback.  Doing a setTimeout after
the change would then cause the notification to fire at the timeout.

Now the Object.observe microtask should be run after the event handler.